// src/generated/models/Pm_performancemeasuresModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_performancemeasurespm_evidenced = {
  0: 'Yes',
  1: 'No'
} as const;
export type Pm_performancemeasurespm_evidenced = keyof typeof Pm_performancemeasurespm_evidenced;
export const Pm_performancemeasuresstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_performancemeasuresstatecode = keyof typeof Pm_performancemeasuresstatecode;
export const Pm_performancemeasuresstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_performancemeasuresstatuscode = keyof typeof Pm_performancemeasuresstatuscode;

export interface Pm_performancemeasuresBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_actualvalue?: number;
  "pm_benefit@odata.bind"?: string;
  pm_benefitcode?: string;
  pm_cumulativeactual?: number;
  pm_cumulativeplanned?: number;
  pm_evidenced?: Pm_performancemeasurespm_evidenced;
  pm_measurename?: string;
  pm_notes?: string;
  pm_performancemeasureid: string;
  pm_plannedvalue?: number;
  pm_reportingperiod?: string;
  pm_variance?: number;
  statecode: Pm_performancemeasuresstatecode;
  statuscode?: Pm_performancemeasuresstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_performancemeasures extends Pm_performancemeasuresBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_benefitname?: string;
  pm_evidencedname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_benefit?: object;
  _pm_benefit_value?: string;
}